Exchange P
The function for exchangeing data has been implemented.
While at the moment only P is being exchanged, the function itself should in theory be able to handle exchange of U and V as well.
This could be applied to exchange the flags a the start to get the proper boundary flags set with cells near the connection of two grids.