Exchange U and V
As mentioned in the previous commit, U and V can be exchanged using the same function that exchanges P.

To do so, the grids for U and V had to be increased in x resp. y direction, because to compute F and G on boundary cells, you need yet another cell beyond that.
Therefore the loops and all indices had to be adjusted once more, but no SegFault is produced, and I think that's a good sign.

After fixing a bug in the translation function (reading the size as a REAL. Yeah sure), output now also works fine,
so the results can once again be visualised using Paraview.