Unify field destruction
It is unnecessary to have two seperate functions to free a double-field vs an integer field.
Hence only one was kept, which also introduces the need to eplicitly type-cast to void* or void**.

Bonus: A memory leak in MPI was closed, by adding a call to MPI_Comm_free at the end of runSimulation.