Test boundary conditions
Add a new test to check the correctness of the velocity fields on the boundary of a general geometry.
In doing so, several errors in the indices of setBCond were discovered (and fixed ofc)

Also, to reduce output in the case of an error, wrap the flag check into in a function and only test the total result.