Test Binary dump
To test the dumping and re-translation of array to and from a binary file, use two methods.
One of them dumps the field to binary and then reads it back, the other one writes directly to VTK.
By comparing the two (using a sys-call to diff), their equality con be determined!